requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
fastapi==0.119.0
filelock==3.20.0
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
pillow==12.0.0
psycopg2==2.9.11
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
python-multipart==0.0.20
setuptools==80.9.0
sniffio==1.3.1
soundfile==0.13.1
SQLAlchemy==2.0.44
starlette==0.48.0
sympy==1.14.0
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
